+++ backend/hexa/webapps/admin.py
@@ -5,13 +5,18 @@ from hexa.webapps.models import SupersetWebapp, Webapp
 
 @admin.register(Webapp)
 class WebappAdmin(admin.ModelAdmin):
-    list_display = ("name", "workspace")
-    list_filter = ("workspace",)
+    list_display = ("name", "workspace", "type", "is_public")
+    list_filter = ("workspace", "type", "is_public")
     search_fields = ("id", "name")
-    fields = ("name", "description", "workspace", "created_by", "url")
-
-    def get_queryset(self, request):
-        return super().get_queryset(request).exclude(type=Webapp.WebappType.SUPERSET)
+    fields = (
+        "name",
+        "description",
+        "workspace",
+        "created_by",
+        "url",
+        "type",
+        "is_public",
+    )
 
 
 class SupersetWebappInline(admin.StackedInline):
@@ -24,8 +29,8 @@ class SupersetWebappInline(admin.StackedInline):
 
 @admin.register(SupersetWebapp)
 class SupersetWebappAdmin(admin.ModelAdmin):
-    list_display = ("name", "workspace", "superset_dashboard")
-    list_filter = ("workspace",)
+    list_display = ("name", "workspace", "is_public", "superset_dashboard")
+    list_filter = ("workspace", "is_public")
     search_fields = ("id", "name")
     raw_id_fields = ("superset_dashboard",)
     fields = (
@@ -34,6 +39,7 @@ class SupersetWebappAdmin(admin.ModelAdmin):
         "workspace",
         "created_by",
         "url",
+        "is_public",
         "superset_dashboard",
     )
     readonly_fields = ("url",)