+++ backend/hexa/workspaces/tests/test_schema/test_workspace.py
@@ -301,39 +301,43 @@ class WorkspaceTest(GraphQLTestCase):
 
     def test_create_workspace_as_org_member(self):
         """Test that any organization member can create workspaces"""
-        self.client.force_login(self.USER_WORKSPACE_EDITOR_ONLY)
-        r = self.run_query(
-            """
-            mutation createWorkspace($input:CreateWorkspaceInput!) {
-                createWorkspace(input: $input) {
-                    success
-                    workspace {
-                        name
-                        description
+        with (
+            patch("hexa.workspaces.models.create_database"),
+            patch("hexa.workspaces.models.load_database_sample_data"),
+        ):
+            self.client.force_login(self.USER_WORKSPACE_EDITOR_ONLY)
+            r = self.run_query(
+                """
+                mutation createWorkspace($input:CreateWorkspaceInput!) {
+                    createWorkspace(input: $input) {
+                        success
+                        workspace {
+                            name
+                            description
+                        }
+                        errors
                     }
-                    errors
                 }
-            }
-            """,
-            {
-                "input": {
-                    "name": "New Workspace by Org Member",
-                    "description": "Created by org member",
-                    "organizationId": str(self.ORGANIZATION.id),
-                }
-            },
-        )
-        self.assertEqual(
-            {
-                "success": True,
-                "errors": [],
-                "workspace": {
-                    "name": "New Workspace by Org Member",
-                    "description": "Created by org member",
+                """,
+                {
+                    "input": {
+                        "name": "New Workspace by Org Member",
+                        "description": "Created by org member",
+                        "organizationId": str(self.ORGANIZATION.id),
+                    }
                 },
-            },
-            r["data"]["createWorkspace"],
-        )
+            )
+            self.assertEqual(
+                {
+                    "success": True,
+                    "errors": [],
+                    "workspace": {
+                        "name": "New Workspace by Org Member",
+                        "description": "Created by org member",
+                    },
+                },
+                r["data"]["createWorkspace"],
+            )
 
     def test_create_workspace_prevent_create(self):
         FeatureFlag.objects.create(